no adapter attached.. again...!!
