+++ app/src/main/java/com/example/umc_hackathon/post/FormListBrand.kt
@@ -1,20 +1,18 @@
-package com.example.umc_hackathon
+package com.example.umc_hackathon.post
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
-import android.view.View.inflate
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.umc_hackathon.databinding.ActivityFormCreateBinding.inflate
-import com.example.umc_hackathon.databinding.ActivityMainBinding
-import com.example.umc_hackathon.databinding.FragmentFormListIdeaBinding
-import com.example.umc_hackathon.databinding.FragmentFormListMarketingBinding
-import com.example.umc_hackathon.databinding.FragmentFormListSocialBinding
+import com.example.umc_hackathon.MySurvey
+import com.example.umc_hackathon.R
+import com.example.umc_hackathon.WaitingSurveyListRAdapter
+import com.example.umc_hackathon.databinding.FragmentFormListBrandBinding
 
-class FormListIdea : Fragment() {
+class FormListBrand : Fragment() {
 
     var modelList = ArrayList<MySurvey>()
     private var linearLayoutManager: RecyclerView.LayoutManager? = null
@@ -22,16 +20,16 @@ class FormListIdea : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
 
-        val binding = FragmentFormListIdeaBinding.inflate(layoutInflater)
+        val binding = FragmentFormListBrandBinding.inflate(layoutInflater)
 
         // 리스트 생성
         for (i in 1..10){
-            val mySurvey = MySurvey(title = "아이디어 $i")
+            val mySurvey = MySurvey(title = "브랜드 $i")
             this.modelList.add(mySurvey)
         }
 
-        val view = inflater!!.inflate(R.layout.fragment_form_list_idea, container, false)
-        val recyclerView: RecyclerView = view.findViewById(R.id.fragment_idea_rv)
+        val view = inflater!!.inflate(R.layout.fragment_form_list_brand, container, false)
+        val recyclerView: RecyclerView = view.findViewById(R.id.fragment_brand_rv)
 
         recyclerAdapter = WaitingSurveyListRAdapter(modelList)
         linearLayoutManager = LinearLayoutManager(activity)
