+++ app/src/main/java/com/example/umc_hackathon/post/FormListRAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.umc_hackathon
+package com.example.umc_hackathon.post
 
 import android.content.Intent
 import android.util.Log
@@ -6,15 +6,17 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import android.widget.Toast
 import androidx.core.content.ContextCompat
 
 import androidx.recyclerview.widget.RecyclerView
-import com.example.umc_hackathon.post.Post
+import com.example.umc_hackathon.FormDetailActivity
+import com.example.umc_hackathon.R
 
 
 class FormListRAdapter(val surveyList: List<Post>): RecyclerView.Adapter<FormListRAdapter.MyViewHolder>() {
 
-    val TAG: String = "<MyRecycleAdapter>"
+    val TAG: String = "<FormListRAdapter>"
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.board_list_item, parent, false)
@@ -22,7 +24,7 @@ class FormListRAdapter(val surveyList: List<Post>): RecyclerView.Adapter<FormLis
             itemView.setOnClickListener {
                 val currentPosition: Int = adapterPosition
                 val survey: Post = surveyList.get(currentPosition)
-//                Toast.makeText(parent.context, "설문조사 제목: ${survey.title}", Toast.LENGTH_SHORT).show()
+                Toast.makeText(parent.context, "설문조사 제목: ${survey.title}", Toast.LENGTH_SHORT).show()
 
                 val intent = Intent(itemView?.context, FormDetailActivity::class.java)
                 intent.putExtra("title", survey.title)
