+++ app/src/main/java/com/example/umc_hackathon/post/FormListSocial.kt
@@ -1,16 +1,15 @@
-package com.example.umc_hackathon
+package com.example.umc_hackathon.post
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
-import android.view.View.inflate
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.umc_hackathon.databinding.ActivityFormCreateBinding.inflate
-import com.example.umc_hackathon.databinding.ActivityMainBinding
-import com.example.umc_hackathon.databinding.FragmentFormListMarketingBinding
+import com.example.umc_hackathon.MySurvey
+import com.example.umc_hackathon.R
+import com.example.umc_hackathon.WaitingSurveyListRAdapter
 import com.example.umc_hackathon.databinding.FragmentFormListSocialBinding
 
 class FormListSocial : Fragment() {
