+++ app/src/main/res/layout/fragment_form_list_brand.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FormListBrand">
+    tools:context=".post.FormListBrand">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fragment_brand_rv"
