+++ app/src/main/res/layout/fragment_form_list_idea.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FormListIdea">
+    tools:context=".post.FormListIdea">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fragment_idea_rv"
