+++ app/src/main/res/layout/fragment_form_list_social.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FormListSocial">
+    tools:context=".post.FormListSocial">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fragment_social_rv"
